# docs/openapi.json
{
...
  "openapi": "3.1.0",
  "info": {
    "title": "PitchFlow API",
    "description": "API documentation for PitchFlow - AI-assisted sponsorship proposal workspace",
    "version": "1.0.0",
    "contact": {
      "name": "Rectoverso Media",
      "email": "support@rectoverso.id"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://pitchflow.vercel.app",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/briefs": {
      "get": {
        "tags": ["Briefs"],
        "summary": "Get all briefs",
        "description": "Retrieve all briefs for the authenticated user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of briefs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Brief" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Briefs"],
        "summary": "Create a new brief",
        "description": "Create a new brief for the authenticated user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["brand_name", "pic_sales", "program"],
                "properties": {
                  "brand_name": { "type": "string", "maxLength": 200 },
                  "pic_sales": { "type": "string", "maxLength": 100 },
                  "program": { "type": "string", "maxLength": 200 },
                  "industry_category": { "type": "string", "maxLength": 100 },
                  "pic_contact": { "type": "string", "maxLength": 100 },
                  "sponsorship_type": { "type": "string", "maxLength": 100 },
                  "objective": { "type": "string", "maxLength": 1000 },
                  "target_audience": { "type": "string", "maxLength": 500 },
                  "period": { "type": "string", "maxLength": 100 },
                  "deadline": { "type": "string", "maxLength": 50 },
                  "budget_range": { "type": "string", "maxLength": 100 },
                  "budget_note": { "type": "string", "maxLength": 500 },
                  "notes": { "type": "string", "maxLength": 2000 },
                  "attachments": {
                    "type": "array",
                    "items": { "type": "string", "format": "uri" },
                    "maxItems": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Brief created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Brief" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Brief": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "brand_name": { "type": "string" },
          "industry_category": { "type": "string" },
          "pic_sales": { "type": "string" },
          "pic_contact": { "type": "string" },
          "program": { "type": "string" },
          "sponsorship_type": { "type": "string" },
          "objective": { "type": "string" },
          "target_audience": { "type": "string" },
          "period": { "type": "string" },
          "deadline": { "type": "string" },
          "budget_range": { "type": "string" },
          "budget_note": { "type": "string" },
          "notes": { "type": "string" },
          "attachments": { "type": "array", "items": { "type": "string" } },
          "status": {
            "type": "string",
            "enum": ["new", "in_review", "in_progress", "completed"]
          },
          "created_at": { "type": "string", "format": "date-time" },
          "created_by": { "type": "string", "format": "uuid" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request - validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": {
              "success": false,
              "error": "Validation failed: brand_name is required"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - authentication required",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": {
              "success": false,
              "error": "Authentication required"
            }
          }
        }
      },
...
    }
  }
}